break out InstanceHistorySerializer
